Move modal files into shared folder
