+++ src/pages/experience/Work.tsx
@@ -1,5 +1,5 @@
 import { ContentItem } from "../components/content item/ContentItem";
-import { Modal } from "../components/Modal";
+import { Modal } from "../components/modal/Modal";
 import { workData } from "../../data/work";
 
 export const Work = () => {
